Catch YAML parse errors when parsing Helm manifests
