+++ source/Calamari.Tests/Calamari.Tests.csproj
@@ -178,6 +178,9 @@
     <None Update="Fixtures\Bash\Scripts\report-kubernetes-manifest-file.sh">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
+    <None Update="KubernetesFixtures\ResourceStatus\assets\invalid-syntax.yaml">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
   <ItemGroup>
     <None Remove="TestResults\**" />
