+++ source/Calamari.Tests/KubernetesFixtures/ResourceStatus/assets/invalid-syntax.yaml
@@ -1,0 +1,34 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  creationTimestamp: null
+  labels:
+    app: nginx
+  name: nginx
+  namespace: test
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: nginx
+  strategy: {}
+  template:
+    metadata:
+      creationTimestamp: null
+      labels:
+        app: nginx
+    spec:
+      containers:
+        - image: nginx
+          name: nginx
+          resources: {}
+status: {}
+---
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: valid-yaml-test
+data:
+  foo.properties: |
+    
+    bar=baz
